+++ launch/autonomous.launch
@@ -27,8 +27,7 @@
 	<node pkg="tricat221" type="autonomous.py" name="autonomous" respawn="true" output="screen"/>
 	<node pkg="tricat221" type="serial_node.py" name="ARDUINO" args="/dev/ttyARDUINO" respawn="true"/>
 
-	<node pkg="rviz" type="rviz" name="rviz_auto" args="-d $(find tricat221)/rviz/rviz_setting1.rviz" />
-	<!--<node pkg="rviz" type="rviz" name="rviz_lidar" args="-d $(find tricat221)/rviz/lidar.rviz" />-->
+	<node pkg="rviz" type="rviz" name="rviz_auto" args="-d $(find tricat221)/rviz/rviz_setting2.rviz" />
 	<node name="rqt" pkg="rqt_graph" type="rqt_graph"/>
 	
 	<!--PARAMETER-->
