+++ launch/simul_autonomous.launch
@@ -4,17 +4,14 @@
 <launch>
 	<node pkg="tricat221" type="lidar_converter.py" name="lidarConverter" respawn="true" output="screen"/>
 	<node pkg="tricat221"  type="heading_calculator.py" name="headingCalculator" respawn="true"/>
-	<node pkg="tricat221" type="autonomous.py" name="autonomous" respawn="true" output="screen">
-		<param name="goal_x" value="3"/>
-		<param name="goal_y" value="3"/>
-	</node>
+	<node pkg="tricat221" type="autonomous.py" name="autonomous" respawn="true" output="screen"/>
 	<node pkg="tricat221" type="moving_avg_filter.py" name="MAF_IMU" output="screen"/>
 	<!--<node pkg="tricat221" type="serial_node.py" name="ARDUINO" args="/dev/ttyARDUINO" respawn="true"/>-->
 
 	<node pkg="rosbag" type="play" name="rosbag_play" required="true" args="$(find tricat221)/etc/rosbag/0429.bag -l --topics /enu_position /imu/data /imu/mag /scan /ublox_gps/fix"/>
 
-	<node pkg="rviz" type="rviz" name="rviz" args="-d $(find tricat221)/rviz/rviz_setting1.rviz" />
-	<node name="rqt" pkg="rqt_graph" type="rqt_graph"/>
+	<node pkg="rviz" type="rviz" name="rviz" args="-d $(find tricat221)/rviz/rviz_setting2.rviz" />
+	<!-- <node name="rqt" pkg="rqt_graph" type="rqt_graph"/> -->
 	
 	<!--PARAMETER-->
 	<rosparam command="load" file="$(find tricat221)/params/lidar_params.yaml"/>
