+++ params/coordinates.yaml
@@ -7,8 +7,8 @@ waypoints: [
             [37.450633, 126.652904, 45.0]
           ]
 
-autonomous_goal: [37.4506473, 126.6528272, 45.0]
-# autonomous_goal: [37.450693, 126.65276, 45.0]
+# autonomous_goal: [37.4506473, 126.6528272, 45.0]
+autonomous_goal: [37.450693, 126.65276, 45.0]
 
 docking_enterence: [37.4511531, 126.6560461, 45.0]
 station1: [37.4511531, 126.6560461, 45.0]
